requirements.txt
fastapi
pydantic
numpy
sentence-transformers
google-generativeai
faiss-cpu
uvicorn
python-dotenv